+++ src/main/java/com/revature/shoply/product/repository/ProductDAO.java
@@ -4,6 +4,7 @@ import com.revature.shoply.models.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
@@ -11,4 +12,6 @@ import java.util.UUID;
 public interface ProductDAO extends JpaRepository<Product, UUID> {
 
     public Optional<Product> findByName(String name);
+
+    public Optional<List<Product>> findByTags_TagName(String name);
 }
