+++ src/main/java/com/revature/shoply/product/service/ProductService.java
@@ -1,12 +1,13 @@
 package com.revature.shoply.product.service;
 
+import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
 import com.revature.shoply.product.DTO.IncomingProductDTO;
 import com.revature.shoply.product.exception.ProductRegistrationException;
 import com.revature.shoply.product.exception.ProductRepositoryException;
-import com.revature.shoply.product.repository.ProductRepository;
+import com.revature.shoply.product.repository.ProductDAO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -18,22 +19,22 @@ import com.revature.shoply.models.Product;
 @Service // service bean entity
 public class ProductService {
 
-    private final ProductRepository productRepository;
+    private final ProductDAO productDAO;
 
     @Autowired
-    public ProductService (ProductRepository productRepository){
-        this.productRepository = productRepository;
+    public ProductService (ProductDAO productRepository){
+        this.productDAO = productRepository;
     }
 
     public Product updateProduct(UUID id, String name, String description, double price) {
-        Optional<Product> optionalProduct = productRepository.findById(id);
+        Optional<Product> optionalProduct = productDAO.findById(id);
 
         if (optionalProduct.isPresent()) {
             Product product = optionalProduct.get();
             product.setName(name);
             product.setDescription(description);
             product.setPrice(price);
-            return productRepository.save(product);
+            return productDAO.save(product);
         } else {
             throw new RuntimeException("Product not found with id: " + id);
         }
@@ -44,15 +45,23 @@ public class ProductService {
         if (productDTO.isValid()) {
 
             // 2. verify product does !exist
-            Optional<Product> existingProduct = productRepository.findByName(productDTO.getName());
+            Optional<Product> existingProduct = productDAO.findByName(productDTO.getName());
             if (!existingProduct.isPresent()) {
 
                 // 3. Convert to Product model instance & return saved product
                 Product product = new Product(null, productDTO.getName(),
                         productDTO.getDescription(), productDTO.getPrice(), productDTO.getQuantity());
-                return productRepository.save(product);
+                return productDAO.save(product);
             } else throw new ProductRepositoryException("A product with that name already exists");
         } else throw new ProductRegistrationException("Invalid product details: mismatched properties");
+    }
+
+    public List<Product> getProductsByTag(String name) {
+        Optional<List<Product>> productsExist = productDAO.findByTags_TagName(name);
+        if(productsExist.isPresent()) {
+            return productsExist.get();
+        }
+        return null;
     }
     
 }
